Add ability to edit scheduled jobs
